+++ packages/admin/src/commands/build.ts
@@ -1,106 +1,54 @@
-import { AdminBuildConfig, build as buildAdmin } from "@medusajs/admin-ui"
-import dotenv from "dotenv"
-import fse from "fs-extra"
-import ora from "ora"
-import { EOL } from "os"
+import { build as adminBuild, clean } from "@medusajs/admin-ui"
 import { resolve } from "path"
-import { loadConfig, reporter, validatePath } from "../utils"
+import { BuildOptions } from "../types"
+import { getPluginPaths, loadConfig } from "../utils"
+import { createBuildManifest } from "../utils/build-manifest"
 
-type BuildArgs = {
-  deployment?: boolean
-  outDir?: string
-  backend?: string
-  include?: string[]
-  includeDist?: string
-  strapi?:string
-  login?:string
-}
+export default async function build({
+  backend,
+  path,
+  outDir,
+  deployment,
+  cmsUrl,
+  loginUrl
+}: BuildOptions) {
+  const {
+    path: configPath,
+    backend: configBackend,
+    outDir: configOutDir,
+  } = loadConfig()
 
-let ENV_FILE_NAME = ""
-switch (process.env.NODE_ENV) {
-  case "production":
-    ENV_FILE_NAME = ".env.production"
-    break
-  case "staging":
-    ENV_FILE_NAME = ".env.staging"
-    break
-  case "test":
-    ENV_FILE_NAME = ".env.test"
-    break
-  case "development":
-  default:
-    ENV_FILE_NAME = ".env"
-    break
-}
+  const plugins = await getPluginPaths()
+  const appDir = process.cwd()
 
-try {
-  dotenv.config({ path: process.cwd() + "/" + ENV_FILE_NAME })
-} catch (e) {
-  reporter.warn(`Failed to load environment variables from ${ENV_FILE_NAME}`)
-}
+  const outDirOption = resolve(appDir, outDir || configOutDir)
+  const pathOption = deployment ? path || "/" : path || configPath
+  const backendOption = deployment
+    ? backend || process.env.MEDUSA_ADMIN_BACKEND_URL
+    : backend || configBackend
 
-export default async function build(args: BuildArgs) {
-  const { deployment, outDir: outDirArg, backend, include, includeDist,strapi,login } = args
-
-  let config: AdminBuildConfig = {}
-
-  if (deployment) {
-    config = {
-      build: {
-        outDir: outDirArg,
-      },
-      globals: {
-        backend: backend || process.env.MEDUSA_BACKEND_URL,
-        strapi: strapi || process.env.STRAPI_URL,
-        login: login || process.env.LOGIN_URL,
-      },
-    }
-  } else {
-    const { path, outDir } = await loadConfig()
-
-    try {
-      validatePath(path)
-    } catch (err) {
-      reporter.panic(err)
-    }
-
-    config = {
-      build: {
-        outDir: outDir,
-      },
-      globals: {
-        base: path,
-        strapi: strapi || process.env.STRAPI_URL,
-        login: login || process.env.LOGIN_URL,
-      },
-    }
-  }
-
-  const time = Date.now()
-  const spinner = ora().start(`Building Admin UI${EOL}`)
-
-  await buildAdmin({
-    ...config,
-  }).catch((err) => {
-    spinner.fail(`Failed to build Admin UI${EOL}`)
-    reporter.panic(err)
+  await clean({
+    appDir: appDir,
+    outDir: outDirOption,
   })
 
-  /**
-   * If we have specified files to include in the build, we copy them
-   * to the build directory.
-   */
-  if (include && include.length > 0) {
-    const dist = outDirArg || resolve(process.cwd(), "build")
+  await adminBuild({
+    appDir: appDir,
+    buildDir: outDirOption,
+    plugins,
+    options: {
+      path: pathOption,
+      backend: backendOption,
+      outDir: outDirOption,
+    },
+  })
 
-    try {
-      for (const filePath of include) {
-        await fse.copy(filePath, resolve(dist, includeDist, filePath))
-      }
-    } catch (err) {
-      reporter.panic(err)
-    }
-  }
-
-  spinner.succeed(`Admin UI build - ${Date.now() - time}ms`)
+  await createBuildManifest(appDir, {
+    outDir: outDir || configOutDir,
+    path: path || configPath,
+    backend: backend || configBackend,
+    deployment,
+    cmsUrl: cmsUrl || process.env.STRAPI_URL,
+    loginUrl: loginUrl || process.env.LOGIN_URL,
+  })
 }