+++ packages/orchestration/src/transaction/transaction-step.ts
@@ -1,15 +1,17 @@
 import { TransactionPayload } from "./distributed-transaction"
+import { DistributedTransaction } from "./distributed-transaction"
 import {
-  TransactionStepsDefinition,
-  TransactionStepStatus,
-  TransactionState,
   TransactionHandlerType,
+  TransactionState,
+  TransactionStepStatus,
+  TransactionStepsDefinition,
 } from "./types"
 
 export type TransactionStepHandler = (
   actionId: string,
   handlerType: TransactionHandlerType,
-  payload: TransactionPayload
+  payload: TransactionPayload,
+  transaction?: DistributedTransaction
 ) => Promise<unknown>
 
 /**