+++ packages/product/src/loaders/container.ts
@@ -52,9 +52,9 @@ export default async ({
   })
 
   if (customRepositories) {
-    await loadCustomRepositories({ customRepositories, container })
+    loadCustomRepositories({ customRepositories, container })
   } else {
-    await loadDefaultRepositories({ container })
+    loadDefaultRepositories({ container })
   }
 }