+++ www/docs/sidebars.js
@@ -99,6 +99,11 @@ module.exports = {
       items: [
         {
           type: "doc",
+          label: "Admin Custom Configuration",
+          id: "admin/configuration",
+        },
+        {
+          type: "doc",
           label: "Admin Widgets",
           id: "admin/widgets",
         },
@@ -106,6 +111,11 @@ module.exports = {
           type: "doc",
           label: "Admin UI Routes",
           id: "admin/routes",
+        },
+        {
+          type: "doc",
+          label: "Admin Setting Pages",
+          id: "admin/setting-pages",
         },
         {
           type: "doc",
@@ -1242,6 +1252,11 @@ module.exports = {
               type: "doc",
               id: "development/entities/create",
               label: "Create an Entity",
+            },
+            {
+              type: "doc",
+              id: "development/entities/repositories",
+              label: "Use a Repository",
             },
             {
               type: "doc",